Completed assignment

+++ Binomial.java
@@ -18,11 +18,11 @@ public class Binomial {
 	}
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
-		int [][] memo = new int[n+1][k+1];
+	public static long binomial(int n, int k) {
+		long [][] memo = new long[n+1][k+1];
 		for (int i = 0; i <= n; i++) {
 			for (int j = 0; j <= k; j++) {
-				memo[i][j] = -1;
+				memo[i][j] = -1L;
 			}
 		}
 		return binomial(n, k, memo);
@@ -32,18 +32,18 @@ public class Binomial {
 		//// It then calls binomial(n, k, memo), which does all the heavy lifiting.
 		//// Replace the following statement with your code.
 
-	private static int binomial(int n, int k, int[][] memo) {
-		if (memo[n][k] != -1) {
+	private static long binomial(int n, int k, long[][] memo) {
+		if (memo[n][k] != -1L) {
 			return memo[n][k];
 		}
 		// Base case
 		if ((k > n)) {
-		   	memo[n][k] = 0; 
+		   	memo[n][k] = 0L; 
 		   	return 0;
 		}
 		// Another base case
 		if (n == 0 || k == 0) {
-		   	memo[n][k] = 1; 
+		   	memo[n][k] = 1L; 
 		   	return 1;
 		}
 		memo[n][k] = binomial(n - 1, k, memo) + binomial(n - 1, k - 1, memo);
